% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutFunction.R
\name{mutantCaller}
\alias{mutantCaller}
\title{Mutant Calling Function for Sanger Sequencing}
\usage{
mutantCaller(test.seq, ref, length, trim5 = 25, trim3 = 25)
}
\arguments{
\item{test.seq}{Input the sequence you wish to test (as a string pointing to a file in your working directory).}

\item{ref}{Input the reference sequencing (you did sequence a control, right?)}

\item{length}{Input the approximate length of your usable sequence as a positive integer.}

\item{trim5}{You can manually adjust the amount of trimming done from the 5' end - defaults to 25.}

\item{trim3}{You can manually adjust the amount of trimming done from the 3' end - defaults to 25.}
}
\description{
This function will allow you to automatically call heterozygous mutations from Sanger sequencing
}
\examples{
mutantCaller(test.seq = './24A_005_JB-T5_JB61_E01.ab1', ref = './24A_001_JB-TM_JB61_A01.ab1', length = 400)
P1               101 ACGACCAGCTCCAGGAGTACGTCCAGCAGACCGTCG--GCCCAGAAACAC    148
                     ||||||||||||||||||||||||||||||||||||  ||||||||||||
S1               101 ACGACCAGCTCCAGGAGTACGTCCAGCAGACCGTCGCAGCCCAGAAACAC    150
This is a two base pair insertion because the secondary sequence is longer than the primary.
}
